+++ 1.21.5/CHANGELOG.md
@@ -4,5 +4,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [v21.5.0-1.21.5] - 2025-04-13
+## [v21.5.0-1.21.5] - 2025-04-14
 - Port to Minecraft 1.21.5
+### Changed
+- The glint overlay for cursed items is now dynamically generated from the vanilla glint overlay
+- This includes a config option for choosing the color of the glint overlay for cursed items